+++ server/tests/unit/data_processors/test_match_data_reader.py
@@ -9,21 +9,23 @@ PROJECT_PATH = os.path.abspath(
 if PROJECT_PATH not in sys.path:
     sys.path.append(PROJECT_PATH)
 
-from app.data_processors import BettingDataReader
+from server.data_processors import MatchDataReader
 
 
-class TestBettingDataReader(TestCase):
+class TestMatchDataReader(TestCase):
     def setUp(self):
-        self.reader = BettingDataReader()
+        self.reader = MatchDataReader()
 
     def test_transform(self):
-        data_frame = self.reader.transform('afl_betting.csv')
+        data_frame = self.reader.transform('ft_match_list.csv')
 
         self.assertIsInstance(data_frame, pd.DataFrame)
 
-        home_columns = data_frame.columns[data_frame.columns.str.match(
-            r'home_')]
-        away_columns = data_frame.columns[data_frame.columns.str.match(
-            r'away_')]
+        home_columns = (
+            data_frame.columns[data_frame.columns.str.match(r'home_')]
+        )
+        away_columns = (
+            data_frame.columns[data_frame.columns.str.match(r'away_')]
+        )
 
         self.assertEqual(len(home_columns), len(away_columns))
