--- tests/feature/test_feature_works.py
@@ -1,2 +1,0 @@
-def test_feature():
-    assert True
